+++ python/tvm/driver/tvmc/composite_target.py
@@ -25,7 +25,7 @@ import tvm.contrib.target.vitis_ai  # pylint: disable=unused-import
 from tvm.relay.op.contrib.arm_compute_lib import partition_for_arm_compute_lib
 from tvm.relay.op.contrib.ethosn import partition_for_ethosn
 from tvm.relay.op.contrib.cmsisnn import partition_for_cmsisnn
-from tvm.relay.backend.contrib.ethosu import partition_for_ethosu
+from tvm.relay.op.contrib.ethosu import partition_for_ethosu
 from tvm.relay.op.contrib.bnns import partition_for_bnns
 from tvm.relay.op.contrib.vitis_ai import partition_for_vitis_ai
